+++ CHANGELOG.md
@@ -8,6 +8,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## Version 4.0.24, 11/20/2024
+
+### Added
+
+N/A
+
+### Removed
+
+N/A
+
+### Changed
+
+Bugfix for an edge case in config reader to handle control character of brackets inside
+an environment variable reference.
+
+e.g. some.key=${ENV_VAR:something/{test1}/{test2}}
+
+---
 ## Version 4.0.23, 11/19/2024
 
 ### Added